+++ packages/editor/src/app/pages/start/components/index.ts
@@ -1,0 +1,3 @@
+export * from './start-new';
+export * from './recent-projects';
+export * from './getting-started';