+++ packages/editor/src/app/pages/workspace/components/pane-details/components/resources.tsx
@@ -20,7 +20,7 @@ export const Resources = () => {
     title: '',
     filename: '',
     ext: '',
-    originalExt: '',
+    sourceExt: '',
     size: 0,
   };
   const [selectedResource, setSelectedResource] =