+++ packages/editor/src/app/services/events/index.ts
@@ -1,0 +1,1 @@
+export * as slide from './slide';