+++ packages/editor/src/app/services/events/slide.ts
@@ -1,0 +1,24 @@
+export const EVENTS = {
+  focus: 'slideFocus',
+};
+
+export const onFocus = (listener) => {
+  document.addEventListener(EVENTS.focus, listener);
+};
+
+export const offFocus = (listener) => {
+  document.removeEventListener(EVENTS.focus, listener);
+};
+
+export const focus = (id: number) => {
+  const event = new CustomEvent(EVENTS.focus, { detail: id });
+
+  document.dispatchEvent(event);
+};
+
+export default {
+  EVENTS,
+  onFocus,
+  offFocus,
+  focus,
+};